+++ packages/util-dev/src/chainSync/index.ts
@@ -98,6 +98,7 @@ const intersect = (events: ChainSyncData['body'], points: PointOrOrigin[]) => {
 };
 
 export enum ChainSyncDataSet {
+  PreviewStakePoolProblem = 'preview-stake-pool-problem.json',
   WithPoolRetirement = 'with-pool-retirement.json',
   WithStakeKeyDeregistration = 'with-stake-key-deregistration.json',
   WithMint = 'with-mint.json',