+++ packages/hardware-ledger/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.8.5](https://github.com/input-output-hk/cardano-js-sdk/compare/@cardano-sdk/hardware-ledger@0.8.4...@cardano-sdk/hardware-ledger@0.8.5) (2024-01-05)
+
+**Note:** Version bump only for package @cardano-sdk/hardware-ledger
+
+## [0.8.4](https://github.com/input-output-hk/cardano-js-sdk/compare/@cardano-sdk/hardware-ledger@0.8.3...@cardano-sdk/hardware-ledger@0.8.4) (2023-12-21)
+
+**Note:** Version bump only for package @cardano-sdk/hardware-ledger
+
 ## [0.8.3](https://github.com/input-output-hk/cardano-js-sdk/compare/@cardano-sdk/hardware-ledger@0.8.2...@cardano-sdk/hardware-ledger@0.8.3) (2023-12-20)
 
 **Note:** Version bump only for package @cardano-sdk/hardware-ledger