+++ packages/hardware-trezor/CHANGELOG.md
@@ -3,6 +3,16 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.4.5](https://github.com/input-output-hk/cardano-js-sdk/compare/@cardano-sdk/hardware-trezor@0.4.4...@cardano-sdk/hardware-trezor@0.4.5) (2024-01-05)
+
+### Features
+
+* **hardware-trezor:** introduce hw trezor multisig signing mode ([5f77781](https://github.com/input-output-hk/cardano-js-sdk/commit/5f777819775f81ae043343de6613fe0ecb5106ce))
+
+## [0.4.4](https://github.com/input-output-hk/cardano-js-sdk/compare/@cardano-sdk/hardware-trezor@0.4.3...@cardano-sdk/hardware-trezor@0.4.4) (2023-12-21)
+
+**Note:** Version bump only for package @cardano-sdk/hardware-trezor
+
 ## [0.4.3](https://github.com/input-output-hk/cardano-js-sdk/compare/@cardano-sdk/hardware-trezor@0.4.2...@cardano-sdk/hardware-trezor@0.4.3) (2023-12-20)
 
 **Note:** Version bump only for package @cardano-sdk/hardware-trezor