+++ packages/input-selection/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.12.12](https://github.com/input-output-hk/cardano-js-sdk/compare/@cardano-sdk/input-selection@0.12.11...@cardano-sdk/input-selection@0.12.12) (2024-01-05)
+
+**Note:** Version bump only for package @cardano-sdk/input-selection
+
+## [0.12.11](https://github.com/input-output-hk/cardano-js-sdk/compare/@cardano-sdk/input-selection@0.12.10...@cardano-sdk/input-selection@0.12.11) (2023-12-21)
+
+**Note:** Version bump only for package @cardano-sdk/input-selection
+
 ## [0.12.10](https://github.com/input-output-hk/cardano-js-sdk/compare/@cardano-sdk/input-selection@0.12.9...@cardano-sdk/input-selection@0.12.10) (2023-12-20)
 
 **Note:** Version bump only for package @cardano-sdk/input-selection