+++ packages/ogmios/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.15.7](https://github.com/input-output-hk/cardano-js-sdk/compare/@cardano-sdk/ogmios@0.15.6...@cardano-sdk/ogmios@0.15.7) (2024-01-05)
+
+**Note:** Version bump only for package @cardano-sdk/ogmios
+
+## [0.15.6](https://github.com/input-output-hk/cardano-js-sdk/compare/@cardano-sdk/ogmios@0.15.5...@cardano-sdk/ogmios@0.15.6) (2023-12-21)
+
+**Note:** Version bump only for package @cardano-sdk/ogmios
+
 ## [0.15.5](https://github.com/input-output-hk/cardano-js-sdk/compare/@cardano-sdk/ogmios@0.15.4...@cardano-sdk/ogmios@0.15.5) (2023-12-20)
 
 **Note:** Version bump only for package @cardano-sdk/ogmios