+++ packages/tx-construction/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.17.2](https://github.com/input-output-hk/cardano-js-sdk/compare/@cardano-sdk/tx-construction@0.17.1...@cardano-sdk/tx-construction@0.17.2) (2024-01-05)
+
+**Note:** Version bump only for package @cardano-sdk/tx-construction
+
+## [0.17.1](https://github.com/input-output-hk/cardano-js-sdk/compare/@cardano-sdk/tx-construction@0.17.0...@cardano-sdk/tx-construction@0.17.1) (2023-12-21)
+
+**Note:** Version bump only for package @cardano-sdk/tx-construction
+
 ## [0.17.0](https://github.com/input-output-hk/cardano-js-sdk/compare/@cardano-sdk/tx-construction@0.16.2...@cardano-sdk/tx-construction@0.17.0) (2023-12-20)
 
 ### ⚠ BREAKING CHANGES