+++ packages/util-dev/CHANGELOG.md
@@ -3,6 +3,16 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.19.5](https://github.com/input-output-hk/cardano-js-sdk/compare/@cardano-sdk/util-dev@0.19.4...@cardano-sdk/util-dev@0.19.5) (2024-01-05)
+
+**Note:** Version bump only for package @cardano-sdk/util-dev
+
+## [0.19.4](https://github.com/input-output-hk/cardano-js-sdk/compare/@cardano-sdk/util-dev@0.19.3...@cardano-sdk/util-dev@0.19.4) (2023-12-21)
+
+### Features
+
+* **util-dev:** add stub data for regression test ([37058d0](https://github.com/input-output-hk/cardano-js-sdk/commit/37058d0f2c2b69e157982bf22330190855feffba))
+
 ## [0.19.3](https://github.com/input-output-hk/cardano-js-sdk/compare/@cardano-sdk/util-dev@0.19.2...@cardano-sdk/util-dev@0.19.3) (2023-12-20)
 
 **Note:** Version bump only for package @cardano-sdk/util-dev