+++ packages/util-rxjs/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.6.8](https://github.com/input-output-hk/cardano-js-sdk/compare/@cardano-sdk/util-rxjs@0.6.7...@cardano-sdk/util-rxjs@0.6.8) (2024-01-05)
+
+**Note:** Version bump only for package @cardano-sdk/util-rxjs
+
+## [0.6.7](https://github.com/input-output-hk/cardano-js-sdk/compare/@cardano-sdk/util-rxjs@0.6.6...@cardano-sdk/util-rxjs@0.6.7) (2023-12-21)
+
+**Note:** Version bump only for package @cardano-sdk/util-rxjs
+
 ## [0.6.6](https://github.com/input-output-hk/cardano-js-sdk/compare/@cardano-sdk/util-rxjs@0.6.5...@cardano-sdk/util-rxjs@0.6.6) (2023-12-20)
 
 **Note:** Version bump only for package @cardano-sdk/util-rxjs